requirement.txt
# Core
python-dotenv
streamlit

# LangChain ecosystem
langchain
langchain-community
langchain-core
langchain-openai
langchain-huggingface

# Vector DB
chromadb

# Embedding / ML
sentence-transformers
torch

# PDF processing
pymupdf
pdfplumber

# Utils
tqdm
numpy

langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
chromadb>=0.5.0
streamlit>=1.35.0
python-dotenv>=1.0.0

sentence-transformers>=3.0.0